Plague Doctor Game Jam 2020 Hack UTD
This is what was uploaded to itch.io at the end of the week for the 2020 Hack UTD game jam. It is buggy and rough around the edges, but as a first foray into game development I am proud of what was accomplished. These files will be added to as the game fleshes out.